+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Smart open is the first engine_type modifier. It now supports compression with all DataConnectors that subclass `FileBasedConnector`. So engine_urls like 'gz+ndjson://myfile.ndjson.gz' are now supported.
+- s3 engine type modifer
+- an example ayeaye.Model that uses s3+gz+csv
 
 ## [0.0.38] - 2023-04-11